Rename: move lib folder in src
