+++ src/provider/QueryProvider.tsx
@@ -4,7 +4,7 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
 import { Toast } from '@/components/common/Toast';
-import { getQueryClient } from '@/utils/query/getQueryClient';
+import { getQueryClient } from '@/lib/query/getQueryClient';
 
 interface QueryProviderProps {
   children: React.ReactNode;
